Simplify the main API
mpn.h now serves as the main public API, exporting only `mpn_state_t`,
`mpn_init()`, and `mpn_parse_line()` exported. For low level access to
handlers, use `mpn_handlers.h`
